Moving dns to dns providers
